🔨 [fix] #346 : guard 버그 수정
