+++ server/src/app.module.ts
@@ -15,7 +15,6 @@ import { MockModule } from "@mock/mock.module";
 import { JwtStrategy } from "@guard/jwt.strategy";
 import { JwtAuthGuard } from "@guard/jwt.guard";
 import { User } from "@user/entities/user.entity";
-import { UserIdGuard } from "@guard/auth.guard";
 import { UploadModule } from "./domain/uploads/upload.module";
 import { typeormConfig } from "./config/typeorm.config";
 
@@ -41,10 +40,6 @@ import { typeormConfig } from "./config/typeorm.config";
     {
       provide: APP_GUARD,
       useClass: JwtAuthGuard,
-    },
-    {
-      provide: APP_GUARD,
-      useClass: UserIdGuard,
     },
     { provide: APP_FILTER, useClass: HttpExceptionFilter },
   ],
