+++ server/src/main.ts
@@ -9,7 +9,6 @@ import express from "express";
 import path from "path";
 import { NestExpressApplication } from "@nestjs/platform-express";
 import { JwtAuthGuard } from "@guard/jwt.guard";
-import { UserIdGuard } from "@guard/auth.guard";
 import { initDatabase } from "./utils/initDB";
 import { AppModule } from "./app.module";
 
@@ -19,7 +18,7 @@ async function bootstrap() {
 
   const app: NestExpressApplication = await NestFactory.create<NestExpressApplication>(AppModule);
   const reflector = app.get(Reflector);
-  app.useGlobalGuards(new JwtAuthGuard(), new UserIdGuard(reflector)); // 전역 user id 검증 가드 적용
+  app.useGlobalGuards(new JwtAuthGuard(reflector)); // 전역 user id 검증 가드 적용
   app.useGlobalFilters(new HttpExceptionFilter()); // 전역 필터 적용
 
   app.use("/user_profiles", express.static(path.join(__dirname, "../user_profiles")));
